requirements1.txt
appdirs==1.4.4
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
dacite==1.6.0
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fake-useragent==0.1.11
feedparser==6.0.8
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
jwcrypto==1.0
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
parse==1.19.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytimeparse==1.1.8
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.8
virtualenv==20.13.1
w3lib==1.22.0
websockets==10.1
wrapt==1.13.3
zipp==3.7.0
